requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
yfinance>=0.2.18
flask>=3.0.0
flask-sqlalchemy>=3.1.0
matplotlib>=3.7.0
